refactor: aggregate hw wallet models
